+++ src/Vector3_TEST.cc
@@ -20,8 +20,8 @@
 #include <numeric>
 #include <sstream>
 
-#include "ignition/math/Vector3.hh"
 #include "ignition/math/Helpers.hh"
+#include "ignition/math/Vector3.hh"
 
 using namespace ignition;
 
@@ -555,4 +555,25 @@ TEST(Vector3dTest, DistToLine)
     math::Vector3d point(0, 2, 0);
     EXPECT_DOUBLE_EQ(point.DistToLine(pointA, pointB), 0);
   }
+}
+
+/////////////////////////////////////////////////
+TEST(Vector3dTest, OperatorStreamOut)
+{
+  math::Vector3d v(0.1234, 1.234, 2.3456);
+  std::ostringstream stream;
+  stream << v;
+  EXPECT_EQ(stream.str(), "0.1234 1.234 2.3456");
+
+  stream.str("");
+  stream << std::setprecision(2) << v;
+  EXPECT_EQ(stream.str(), "0.12 1.2 2.3");
+
+  stream.str("");
+  stream << std::setprecision(3) << v;
+  EXPECT_EQ(stream.str(), "0.123 1.23 2.35");
+
+  stream.str("");
+  stream << std::setprecision(1) << std::fixed << v;
+  EXPECT_EQ(stream.str(), "0.1 1.2 2.3");
 }