+++ src/Vector4_TEST.cc
@@ -379,9 +379,21 @@ TEST(Vector4dTest, Sub)
 /////////////////////////////////////////////////
 TEST(Vector4dTest, OperatorStreamOut)
 {
-  math::Vector4d v(0.1, 1.2, 2.3, 0.0);
+  math::Vector4d v(0.1234, 1.234, 2.3456, 0.0);
   std::ostringstream stream;
   stream << v;
+  EXPECT_EQ(stream.str(), "0.1234 1.234 2.3456 0");
+
+  stream.str("");
+  stream << std::setprecision(2) << v;
+  EXPECT_EQ(stream.str(), "0.12 1.2 2.3 0");
+
+  stream.str("");
+  stream << std::setprecision(3) << v;
+  EXPECT_EQ(stream.str(), "0.123 1.23 2.35 0");
+
+  stream.str("");
+  stream << std::setprecision(1) << std::fixed << v;
   EXPECT_EQ(stream.str(), "0.1 1.2 2.3 0");
 }